fix(parser): read/write IfcRelNests for IfcWorkPlan nesting IfcWorkSchedule
IfcWorkPlan groups IfcWorkSchedules via IfcRelNests. Neither direction was
handled: the extractor's IfcRelNests pass only ever resolved a nesting
parent through the task table (a WorkPlan RelatingObject was silently
dropped, with nowhere on WorkScheduleInfo for the data to land even if it
had matched), and the serializer never emitted this relation at all — only
task/subtask IFCRELNESTS, IFCRELASSIGNSTOCONTROL, IFCRELASSIGNSTOPROCESS and
IFCRELSEQUENCE.

Confirmed against buildingSMART's own IFC4 spec reference file for IfcTask,
which uses exactly this pattern (a work plan nesting a work schedule via
IfcRelNests). ifc-lite's schedule-roundtrip suite couldn't catch this: it
round-trips the writer through the reader, so a relation the writer never
emitted can't appear as a mismatch there.

WorkScheduleInfo gains childScheduleGlobalIds (on a WorkPlan) and
parentPlanGlobalId (on a WorkSchedule), populated by a new extractor pass
and written back by a matching serializer section.

Per-relation verdict for IfcRelNests in the schedule extractor: task/subtask
hierarchy was already correct; WorkPlan -> WorkSchedule was the gap fixed
here. Also noted, not fixed here (separate relation, separate issue): the
SDK's assignSchedulesToWorkPlan groups via IfcRelAssignsToControl instead,
and that grouping isn't read back either, for the same taskByExpressId-only
shape in the extractor's Pass 5.

Split both touched files rather than raising their module-size budgets, since
the growth was entirely from this change: schedule-extractor.ts split into
schedule-extractor.ts (cross-entity orchestration, the multi-pass walk) and
schedule-types.ts (STEP attribute-index layouts, the record types, and the
small pure/single-entity helpers those passes call) — same shape as this
package's existing spatial-hierarchy-canonical-parent.ts /
spatial-hierarchy-builder.ts split. schedule-serializer.ts split the same
way into schedule-serializer.ts (orchestration) and
schedule-serializer-builders.ts (STEP string encoding + per-entity line
builders). Both new files land at 199 and 357 lines — well under the
400-line house rule — so neither needs an allowlist row, and
schedule-extractor.ts itself dropped from 594 (its old budget) to 348, so
its allowlist row was deleted rather than lowered, per the gate's own
"total must trend down" rule.

Fixes #4329.

Closes an unlabelled issue, so this PR carries `unqueued` per the issue-queue
gate rather than a `ready` label I'm not positioned to self-apply.
